+++ starks/stark.py
@@ -1,91 +1,342 @@
 import time
+from starks.merkle_tree import merkelize, mk_branch, verify_branch, blake
+from starks.compression import compress_fri, decompress_fri, compress_branches, decompress_branches, bin_length
+from starks.poly_utils import PrimeField
+from starks.fft import fft
+from starks.fri import prove_low_degree, verify_low_degree_proof
+from starks.utils import get_power_cycle, get_pseudorandom_indices, is_a_power_of_2
 
-modulus = 2**256 - 351 * 2**32 + 1
+modulus = 2**256 - 2**32 * 351 + 1
+f = PrimeField(modulus)
+nonresidue = 7
 
-#class PrimeField():
-#    """A simple class implementing a prime field."""
-#
-#    def __init__(self, modulus):
-#        # Quick primality test
-#        assert pow(2, modulus, modulus) == 2
-#        self.modulus = modulus
-#
-#    def add(self, x, y):
-#        return (x+y) % self.modulus
-#
-#    def sub(self, x, y):
-#        return (x-y) % self.modulus
-#
-#    def mul(self, x, y):
-#        return (x*y) % self.modulus
-#
-#    def inv(self, a):
-#        """Modular inverse using the extended Euclidean algorithm"""
-#        if a == 0:
-#            return 0
-#        lm, hm = 1, 0
-#        low, high = a % self.modulus, self.modulus
-#        while low > 1:
-#            r = high//low
-#            nm, new = hm-lm*r, high-low*r
-#            lm, low, hm, high = nm, new, lm, low
-#        return lm % self.modulus
-#
-#    def multi_inv(self, values):
-#        partials = [1]
-#        for i in range(len(values)):
-#            partials.append(self.mul(partials[-1], values[i] or 1))
-#        inv = self.inv(partials[-1])
-#        outputs = [0] * len(values)
-#        for i in range(len(values), 0, -1):
-#            outputs[i-1] = self.mul(partials[i-1], inv) if values[i-1] else 0
-#            inv = self.mul(inv, values[i-1] or 1)
-#        return outputs
-#
-#    # Evaluate a polynomial at a point
-#    def eval_poly_at(self, p, x):
-#        y = 0
-#        power_of_x = 1
-#        for i, p_coeff in enumerate(p):
-#            y += power_of_x * p_coeff
-#            power_of_x = (power_of_x * x) % self.modulus
-#        return y % self.modulus
-#
-#    # Build a polynomial that returns 0 at all specified xs
-#    def zpoly(self, xs):
-#        root = [1]
-#        for x in xs:
-#            root.insert(0, 0)
-#            for j in range(len(root)-1):
-#                root[j] -= root[j+1] * x
-#        return [x % self.modulus for x in root]
-#
-#    def lagrange_interp(self, xs, ys):
-#        # Generate master numerator polynomial, eg. (x - x1) * (x - x2) * ... * (x - xn)
-#        root = self.zpoly(xs)
-#
-#        # Generate per-value numerator polynomials, eg. for x=x2,
-#        # (x - x1) * (x - x3) * ... * (x - xn), by dividing the master
-#        # polynomial back by each x coordinate
-#        nums = [self.div_polys(root, [-x, 1]) for x in xs]
-#
-#        # Generate denominators by evaluating numerator polys at each x
-#        denoms = [self.eval_poly_at(nums[i], xs[i]) for i in range(len(xs))]
-#        invdenoms = self.multi_inv(denoms)
-#
-#        # Generate output polynomial, which is the sum of the per-value numerator
-#        # polynomials rescaled to have the right y values
-#        b = [0 for y in ys]
-#        for i in range(len(xs)):
-#            yslice = self.mul(ys[i], invdenoms[i])
-#            for j in range(len(ys)):
-#                if nums[i][j] and ys[i]:
-#                    b[j] += nums[i][j] * yslice
-#        return [x % self.modulus for x in b]
+# Number of bits of security in Merkle-tree check
+spot_check_security_factor = 80
+# TODO(rbharath): Is this the Galois extension degree?
+# I think this is a security factor that gives us more control
+# over the security level of the computation, but not sure.
+extension_factor = 8
 
-#def mimc(inp, steps, round_constants):
-#    start_time = time.time()
-#    for i in range(steps-1):
-#        inp = (inp**3 + round_constants[i % len(round_constants)]) % modulus
-#    print("MIMC computed in %.4f sec" % (time.time() - start_time))
-#    return inp
+
+# TODO(rbharath): Wait, does Vitalik's blog post claim that
+# the verifier complexity is linear; Nah looks like t*log(t)
+# is optimal. Verifier complexity is O(log**2(t)) which should
+# be pretty small even for very large computations.
+# NOTE(rbharath): These starks here are not zero-knowledge I
+# think. Will need to be added onto library later.
+def mimc(inp, steps, round_constants):
+  """Compute a MIMC permutation for some number of steps"""
+  start_time = time.time()
+  for i in range(steps - 1):
+    inp = (inp**3 + round_constants[i % len(round_constants)]) % modulus
+  print("MIMC computed in %.4f sec" % (time.time() - start_time))
+  return inp
+
+def get_computational_trace(inp, steps, computational_step):
+  """Get the computational trace for the STARK.
+
+  This function is a first step towards refactoring this code
+  so it can generate STARKs for different computations. For
+  now, this only works for MiMC
+
+  Parameters
+  ----------
+  inp: Int
+    The input for the computation
+  """
+  computational_trace = [inp]
+  for i in range(steps - 1):
+    computational_trace.append(computational_step(computational_trace[-1], i))
+  output = computational_trace[-1]
+  print('Done generating computational trace')
+  return computational_trace, output
+
+def construct_constraint_polynomial(steps, round_constants, G1, G2, precision, p_evaluations, update_fn):
+  """Construct the constraint polynomial for the given tape.
+
+  This function constructs a constraint polynomial for the
+  given computational tape. For now, this function only works
+  with MiMC.
+  """
+  skips2 = steps // len(round_constants)
+  constants_mini_polynomial = fft(
+      round_constants, modulus, f.exp(G1, skips2), inv=True)
+  constants_mini_extension = fft(constants_mini_polynomial, modulus,
+                                 f.exp(G2, skips2))
+  #print("extension_factor")
+  #print(extension_factor)
+  #print("precision, skips2")
+  #print(precision, skips2)
+  #print("len(constants_mini_extension)")
+  #print(len(constants_mini_extension))
+  #print("constants_mini_extension[0]")
+  #print(constants_mini_extension[0])
+  #print("constants_mini_extension[8]")
+  #print(constants_mini_extension[8])
+  #print("constants_mini_extension[64]")
+  #print(constants_mini_extension[64])
+  #print("constants_mini_extension[63]")
+  #print(constants_mini_extension[63])
+  #print("constants_mini_extension[128]")
+  #print(constants_mini_extension[128])
+  #assert 0 == 1
+  print(
+      'Converted round constants into a polynomial and low-degree extended it')
+
+  # Create the composed polynomial such that
+  # C(P(x), P(g1*x), K(x)) = P(g1*x) - P(x)**3 - K(x)
+  # here K(x) is the round constants.
+  # TODO(rbharath): Why does this structure make sense? Oh, I
+  # think this is the forward loop in MiMC. g1*x is the next
+  # iteration and should equal the MiMC pass from previous.
+  # For a deep network, would be the sigma(wx+b) I think.
+  c_of_p_evaluations = [
+      (p_evaluations[
+          (i + extension_factor) % precision] - update_fn(f, p_evaluations[i], i) -
+       constants_mini_extension[i % len(constants_mini_extension)]) % modulus
+      for i in range(precision)
+  ]
+  print('Computed C(P, K) polynomial')
+  return c_of_p_evaluations
+
+def compute_remainder_polynomial(xs, precision, steps, last_step_position, c_of_p_evaluations):
+  """Computes the remainder polynomial for the STARK.
+  
+  Compute D(x) = C(P(x), P(g1*x), K(x)) / Z(x)
+  Z(x) = (x^steps - 1) / (x - x_atlast_step)
+  TODO(rbharath): I think this is supposed to equal 
+  Z(x) = (x - 1)(x-2)...(x-(steps_1)). How are these equal?
+  """
+  z_num_evaluations = [
+      xs[(i * steps) % precision] - 1 for i in range(precision)
+  ]
+  z_num_inv = f.multi_inv(z_num_evaluations)
+  # (x_i - x_{step-1}) list
+  z_den_evaluations = [xs[i] - last_step_position for i in range(precision)]
+  d_evaluations = [
+      cp * zd * zni % modulus
+      for cp, zd, zni in zip(c_of_p_evaluations, z_den_evaluations, z_num_inv)
+  ]
+  print('Computed D polynomial')
+  return d_evaluations
+
+def compute_boundary_polynomial(xs, last_step_position, inp, output, p_evaluations):
+  """Polynomial encoding boundary constraints on tape.
+  
+  Compute interpolant of ((1, input), (x_atlast_step, output))
+  """
+  # TODO(rbharath): Why does this interpolant make sense?
+  interpolant = f.lagrange_interp_2([1, last_step_position], [inp, output])
+  i_evaluations = [f.eval_poly_at(interpolant, x) for x in xs]
+
+  zeropoly2 = f.mul_polys([-1, 1], [-last_step_position, 1])
+  inv_z2_evaluations = f.multi_inv([f.eval_poly_at(zeropoly2, x) for x in xs])
+
+  # B = (P - I) / Z2
+  b_evaluations = [
+      ((p - i) * invq) % modulus
+      for p, i, invq in zip(p_evaluations, i_evaluations, inv_z2_evaluations)
+  ]
+  print('Computed B polynomial')
+  return b_evaluations
+
+
+def compute_pseudorandom_linear_combination(mtree, G2, steps, precision, d_evaluations, p_evaluations, b_evaluations):
+  """Computes a psuedorandom linear combination of polys
+
+  A FRI proofs of low degree for a polynomial takes space. There are multiple polynomials (constraint, tape, reminder, degree) used in a STARK. This function combines these into a single polynomial so that only one FRI proofs needs to be generated for all of them. THe chances of a collision are low.
+  """
+  # Based on the hashes of P, D and B, we select a random
+  # linear combination of P * x^steps, P, B * x^steps, B and
+  # D, and prove the low-degreeness of that, instead of
+  # proving the low-degreeness of P, B and D separately
+  k1 = int.from_bytes(blake(mtree[1] + b'\x01'), 'big')
+  k2 = int.from_bytes(blake(mtree[1] + b'\x02'), 'big')
+  k3 = int.from_bytes(blake(mtree[1] + b'\x03'), 'big')
+  k4 = int.from_bytes(blake(mtree[1] + b'\x04'), 'big')
+
+  # Compute the linear combination. We don't even both
+  # calculating it in coefficient form; we just compute the
+  # evaluations
+  G2_to_the_steps = f.exp(G2, steps)
+  powers = [1]
+  for i in range(1, precision):
+    powers.append(powers[-1] * G2_to_the_steps % modulus)
+
+  l_evaluations = [(d_evaluations[i] + p_evaluations[i] * k1 +
+                    p_evaluations[i] * k2 * powers[i] + b_evaluations[i] * k3 +
+                    b_evaluations[i] * powers[i] * k4) % modulus
+                   for i in range(precision)]
+
+  print('Computed random linear combination')
+  return l_evaluations
+
+def compute_merkle_spot_checks(mtree, l_mtree, precision, skips):
+  """Computes pseudorandom spot checks of Merkle tree."""
+  # Do some spot checks of the Merkle tree at pseudo-random
+  # coordinates, excluding multiples of `extension_factor`
+  branches = []
+  samples = spot_check_security_factor
+  positions = get_pseudorandom_indices(
+      l_mtree[1], precision, samples, exclude_multiples_of=extension_factor)
+  for pos in positions:
+    branches.append(mk_branch(mtree, pos))
+    branches.append(mk_branch(mtree, (pos + skips) % precision))
+    branches.append(mk_branch(l_mtree, pos))
+  print('Computed %d spot checks' % samples)
+  return branches
+
+
+# NOTE(rbharath): If you run a deep learning model on a GPU,
+# you can add a trace-log which can be exited from the GPU
+# without much effort. This trace log can be passed along to
+# the STARK prover off-line.
+
+# TODO(rbharath): I think the general strategy is to create a
+# "comptutational "trace" of the workload, then constract a
+# polynomial constraint with necessary linkage. Easier for
+# regular workloads. This is also called a "computation tape"
+
+def mk_proof(inp, steps, round_constants, step_fn, update_fn):
+  """Generate a STARK for a MIMC calculation"""
+  start_time = time.time()
+  # Some constraints to make our job easier
+  # TODO: How can round_constants be factored out here
+  assert steps <= 2**32 // extension_factor
+  assert is_a_power_of_2(steps) and is_a_power_of_2(len(round_constants))
+  assert len(round_constants) <= steps
+
+  precision = steps * extension_factor
+
+  # Root of unity such that x^precision=1
+  G2 = f.exp(7, (modulus - 1) // precision)
+
+  # Root of unity such that x^steps=1
+  skips = precision // steps
+  assert skips == extension_factor
+  G1 = f.exp(G2, skips)
+
+  # Powers of the higher-order root of unity
+  xs = get_power_cycle(G2, modulus)
+  last_step_position = xs[(steps - 1) * extension_factor]
+
+  # computational_trace is a tape of computation values. (Put
+  # another way, a list of partial values the computation
+  # takes on).
+  computational_trace, output = get_computational_trace(inp, steps, step_fn)
+
+  # Interpolate the computational trace into a polynomial P,
+  # with each step along a successive power of G1
+  computational_trace_polynomial = fft(
+      computational_trace, modulus, G1, inv=True)
+  p_evaluations = fft(computational_trace_polynomial, modulus, G2)
+  print(
+      'Converted computational steps into a polynomial and low-degree extended it'
+  )
+
+  # Construct the constraint polynomial (represented as a list
+  # of point evaluations)
+  c_of_p_evaluations = construct_constraint_polynomial(steps, round_constants, G1, G2, precision, p_evaluations, update_fn)
+
+  d_evaluations = compute_remainder_polynomial(xs, precision, steps, last_step_position, c_of_p_evaluations)
+
+  b_evaluations = compute_boundary_polynomial(xs, last_step_position, inp, output, p_evaluations)
+
+  # Compute their Merkle root
+  mtree = merkelize([
+      pval.to_bytes(32, 'big') + dval.to_bytes(32, 'big') + bval.to_bytes(
+          32, 'big')
+      for pval, dval, bval in zip(p_evaluations, d_evaluations, b_evaluations)
+  ])
+  print('Computed hash root')
+
+  l_evaluations = compute_pseudorandom_linear_combination(mtree, G2, steps, precision, d_evaluations, p_evaluations, b_evaluations)
+  l_mtree = merkelize(l_evaluations)
+
+  branches = compute_merkle_spot_checks(mtree, l_mtree, precision, skips)
+
+  # Return the Merkle roots of P and D, the spot check Merkle
+  # proofs, and low-degree proofs of P and D
+  o = [
+      mtree[1], l_mtree[1], branches,
+      prove_low_degree(
+          l_evaluations,
+          G2,
+          steps * 2,
+          modulus,
+          exclude_multiples_of=extension_factor)
+  ]
+  print("STARK computed in %.4f sec" % (time.time() - start_time))
+  return o
+
+
+def verify_proof(inp, steps, round_constants, output, proof, transition_constraint):
+  """Verifies a STARK"""
+  m_root, l_root, branches, fri_proof = proof
+  start_time = time.time()
+  assert steps <= 2**32 // extension_factor
+  assert is_a_power_of_2(steps) and is_a_power_of_2(len(round_constants))
+  assert len(round_constants) <= steps
+
+  precision = steps * extension_factor
+
+  # Get (steps)th root of unity
+  G2 = f.exp(7, (modulus - 1) // precision)
+  skips = precision // steps
+
+  # Gets the polynomial representing the round constants
+  skips2 = steps // len(round_constants)
+  constants_mini_polynomial = fft(
+      round_constants, modulus, f.exp(G2, extension_factor * skips2), inv=True)
+
+  # Verifies the low-degree proofs
+  assert verify_low_degree_proof(
+      l_root,
+      G2,
+      fri_proof,
+      steps * 2,
+      modulus,
+      exclude_multiples_of=extension_factor)
+
+  # Performs the spot checks
+  k1 = int.from_bytes(blake(m_root + b'\x01'), 'big')
+  k2 = int.from_bytes(blake(m_root + b'\x02'), 'big')
+  k3 = int.from_bytes(blake(m_root + b'\x03'), 'big')
+  k4 = int.from_bytes(blake(m_root + b'\x04'), 'big')
+  samples = spot_check_security_factor
+  positions = get_pseudorandom_indices(
+      l_root, precision, samples, exclude_multiples_of=extension_factor)
+  last_step_position = f.exp(G2, (steps - 1) * skips)
+  for i, pos in enumerate(positions):
+    x = f.exp(G2, pos)
+    x_to_the_steps = f.exp(x, steps)
+    mbranch1 = verify_branch(m_root, pos, branches[i * 3])
+    mbranch2 = verify_branch(m_root, (pos + skips) % precision,
+                             branches[i * 3 + 1])
+    l_of_x = verify_branch(l_root, pos, branches[i * 3 + 2], output_as_int=True)
+
+    p_of_x = int.from_bytes(mbranch1[:32], 'big')
+    p_of_g1x = int.from_bytes(mbranch2[:32], 'big')
+    d_of_x = int.from_bytes(mbranch1[32:64], 'big')
+    b_of_x = int.from_bytes(mbranch1[64:], 'big')
+
+    zvalue = f.div(f.exp(x, steps) - 1, x - last_step_position)
+    k_of_x = f.eval_poly_at(constants_mini_polynomial, f.exp(x, skips2))
+
+    # Check transition constraints C(P(x)) = Z(x) * D(x)
+    #assert (p_of_g1x - p_of_x**3 - k_of_x - zvalue * d_of_x) % modulus == 0
+    assert (p_of_g1x - transition_constraint(p_of_x) - k_of_x - zvalue * d_of_x) % modulus == 0
+
+    # Check boundary constraints B(x) * Q(x) + I(x) = P(x)
+    interpolant = f.lagrange_interp_2([1, last_step_position], [inp, output])
+    zeropoly2 = f.mul_polys([-1, 1], [-last_step_position, 1])
+    assert (p_of_x - b_of_x * f.eval_poly_at(zeropoly2, x) - f.eval_poly_at(
+        interpolant, x)) % modulus == 0
+
+    # Check correctness of the linear combination
+    assert (l_of_x - d_of_x - k1 * p_of_x - k2 * p_of_x * x_to_the_steps -
+            k3 * b_of_x - k4 * b_of_x * x_to_the_steps) % modulus == 0
+
+  print('Verified %d consistency checks' % spot_check_security_factor)
+  print('Verified STARK in %.4f sec' % (time.time() - start_time))
+  return True
